[exotica_examples] Update unittest URDFs to fix failures with latest MoveIt
The latest version of MoveIt correctly distinguishes visual and
collision links. For setting up the unittests, we incorrectly only used
the visual-tag. Formerly, MoveIt interpreted the visual tag also as a
collision tag if no collision tag was specified. This has now been
changed. As a result, we need to fix our URDFs that made incorrect use
of that upstream bug.