[exotica_python] Fix #719: Adding unknown initializer properties
This commit fixes #719 where unknown properties of a base initializer
were not added to the property (were ignored). This prevented TaskMaps
or other components that require explicitly casting a base/abstract
initializer to a derived class to work.

For instance: SphereCollision uses an augmented Initializer "Sphere"
which derives of "Frame" to specify spheres within the generic
"EndEffector" list of initializers. Without this commit, exotica_python
would drop the Radius and Group properties since they are not part of
the Frame initializer. However, the taskmap does require them so this
commit will make sure they are properly added even though not
technically part of the Frame initializer definition. It will print
warnings though to alert users that this might be a potential issue.