[exotica_core[ Add test_helpers.h to remove problematic definition of ColoredPrintf
The redefinition of color enums for ColoredPrintf within googletest was
not portable between different versions of googletest. This lead for
instance to failures on Noetic which we could only partially fix with
version checks (e.g. it was still failing on Debian Buster). With this
commit we are adding a new header to be used in tests that defines
TEST_COUT using ANSI color codes and does not require
googletest-internal definitions. This fixes the observed problem in
Buster and further reduces duplication in our testing code :-).